agent-orchestrator/requirements.txt
langchain-core>=0.3
langchain-anthropic>=0.3
langchain-openai>=0.3
langchain-ollama>=0.3
fastapi>=0.115
uvicorn>=0.34
pydantic-settings>=2.0
pyyaml>=6.0
sse-starlette>=2.0
neo4j>=5.0
